Rename script language class names.
Fixes #123.
