+++ src/main/scala/com/exasol/cloudetl/scriptclasses/ImportMetadata.scala
@@ -2,30 +2,14 @@ package com.exasol.cloudetl.scriptclasses
 
 import com.exasol.ExaIterator
 import com.exasol.ExaMetadata
-import com.exasol.cloudetl.bucket.Bucket
-import com.exasol.cloudetl.storage.StorageProperties
 
 import com.typesafe.scalalogging.LazyLogging
 
 object ImportMetadata extends LazyLogging {
 
-  def run(metadata: ExaMetadata, iterator: ExaIterator): Unit = {
-    val bucketPath = iterator.getString(0)
-    val parallelism = iterator.getInteger(2)
-    logger.info(
-      s"Reading metadata from bucket path: $bucketPath "
-        + s"with parallelism: ${parallelism.toString}"
+  def run(metadata: ExaMetadata, iterator: ExaIterator): Unit =
+    throw new ImportScriptClassException(
+      "This script class is deprecated. Please use the FilesMetadataReader " +
+        "class name. You can check the user guide for updated deployment scripts."
     )
-
-    val storageProperties = StorageProperties(iterator.getString(1), metadata)
-    val bucket = Bucket(storageProperties)
-    val paths = bucket.getPaths()
-    logger.info(s"Total number of files: ${paths.size} in bucket path: $bucketPath")
-
-    paths.zipWithIndex.foreach {
-      case (filename, idx) =>
-        iterator.emit(filename.toString, s"${idx % parallelism}")
-    }
-  }
-
 }
