+++ src/main/scala/com/exasol/cloudetl/scriptclasses/ImportPath.scala
@@ -1,39 +1,13 @@
 package com.exasol.cloudetl.scriptclasses
 
-import scala.collection.JavaConverters._
-
 import com.exasol.ExaImportSpecification
 import com.exasol.ExaMetadata
-import com.exasol.cloudetl.bucket.Bucket
-import com.exasol.cloudetl.storage.StorageProperties
 
 object ImportPath {
 
-  def generateSqlForImportSpec(
-    metadata: ExaMetadata,
-    importSpec: ExaImportSpecification
-  ): String = {
-    val storageProperties = StorageProperties(importSpec.getParameters.asScala.toMap)
-    val bucket = Bucket(storageProperties)
-    bucket.validate()
-
-    val scriptSchema = metadata.getScriptSchema
-    val bucketPath = bucket.bucketPath
-    val parallelism = storageProperties.getParallelism("nproc()")
-    val storagePropertiesAsString = storageProperties.mkString()
-
-    s"""SELECT
-       |  $scriptSchema.IMPORT_FILES(
-       |    '$bucketPath', '$storagePropertiesAsString', filename
-       |)
-       |FROM (
-       |  SELECT $scriptSchema.IMPORT_METADATA(
-       |    '$bucketPath', '$storagePropertiesAsString', $parallelism
-       |  )
-       |)
-       |GROUP BY
-       |  partition_index;
-       |""".stripMargin
-  }
-
+  def generateSqlForImportSpec(meta: ExaMetadata, spec: ExaImportSpecification): String =
+    throw new ImportScriptClassException(
+      "This script class is deprecated. Please use the FilesImportQueryGenerator " +
+        "class name. You can check the user guide for updated deployment scripts."
+    )
 }
