+++ src/main/scala/com/exasol/cloudetl/scriptclasses/ImportScriptClassException.scala
@@ -1,0 +1,6 @@
+package com.exasol.cloudetl.scriptclasses
+
+class ImportScriptClassException(val message: String, val cause: Throwable)
+    extends RuntimeException(message, cause) {
+  def this(message: String) = this(message, null)
+}
