+++ src/test/scala/com/exasol/cloudetl/scriptclasses/ImportMetadataTest.scala
@@ -2,31 +2,14 @@ package com.exasol.cloudetl.scriptclasses
 
 import com.exasol.ExaMetadata
 
-import org.mockito.ArgumentMatchers.anyString
-import org.mockito.Mockito._
-
 class ImportMetadataTest extends StorageTest {
 
-  test("run returns the list of file names") {
-    val properties = Map(
-      "BUCKET_PATH" -> testResourceParquetPath,
-      "DATA_FORMAT" -> "PARQUET"
-    )
-    val expectedParquetFiles = Map(
-      s"$testResourceDir/import/parquet/sales_positions1.snappy.parquet" -> "0",
-      s"$testResourceDir/import/parquet/sales_positions2.snappy.parquet" -> "1",
-      s"$testResourceDir/import/parquet/sales_positions_small.snappy.parquet" -> "0"
-    )
-
-    val iter = mockExasolIterator(properties)
-    when(iter.getInteger(2)).thenReturn(2)
-
-    ImportMetadata.run(mock[ExaMetadata], iter)
-    verify(iter, times(3)).emit(anyString(), anyString())
-    expectedParquetFiles.foreach {
-      case (filename, partitionId) =>
-        verify(iter, times(1)).emit(filename, partitionId)
+  test("run throws with message to the new scriptclass name") {
+    val properties = Map("BUCKET_PATH" -> "empty", "DATA_FORMAT" -> "avro")
+    val thrown = intercept[ImportScriptClassException] {
+      ImportMetadata.run(mock[ExaMetadata], mockExasolIterator(properties))
     }
+    assert(thrown.getMessage().contains("Please use the FilesMetadataReader"))
   }
 
 }
